@@ -2,6 +2,6 @@ module modernc.org/memory
 
 // +heroku goVersion go1.14
 
-go 1.15
+go 1.16
 
 require modernc.org/mathutil v1.4.1
