@@ -1,13 +1,32 @@
 module modernc.org/sqlite
 
-go 1.16
+go 1.17
 
 require (
-	github.com/mattn/go-sqlite3 v1.14.9
+	github.com/mattn/go-sqlite3 v1.14.12
 	golang.org/x/sys v0.0.0-20211007075335-d3039528d8ac
-	modernc.org/ccgo/v3 v3.12.95
-	modernc.org/libc v1.11.104
+	modernc.org/ccgo/v3 v3.16.6
+	modernc.org/libc v1.16.7
 	modernc.org/mathutil v1.4.1
-	modernc.org/tcl v1.9.2
-	modernc.org/z v1.2.20
+	modernc.org/tcl v1.13.1
 )
+
+require (
+	github.com/google/uuid v1.3.0 // indirect
+	github.com/kballard/go-shellquote v0.0.0-20180428030007-95032a82bc51 // indirect
+	github.com/mattn/go-isatty v0.0.12 // indirect
+	github.com/remyoudompheng/bigfft v0.0.0-20200410134404-eec4a21b6bb0 // indirect
+	golang.org/x/mod v0.3.0 // indirect
+	golang.org/x/tools v0.0.0-20201124115921-2c860bdd6e78 // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	lukechampine.com/uint128 v1.1.1 // indirect
+	modernc.org/cc/v3 v3.36.0 // indirect
+	modernc.org/httpfs v1.0.6 // indirect
+	modernc.org/memory v1.1.1 // indirect
+	modernc.org/opt v0.1.1 // indirect
+	modernc.org/strutil v1.1.1 // indirect
+	modernc.org/token v1.0.0 // indirect
+	modernc.org/z v1.5.1 // indirect
+)
+
+retract [v1.16.0, v1.17.2] // https://gitlab.com/cznic/sqlite/-/issues/100
