@@ -57,7 +57,7 @@ func NextWaitInterval(lastWaitInterval time.Duration, err error) time.Duration {
 	}
 
 	// Add some jitter to avoid unnecessary collision between competing other instances.
-	nextWaitInterval += time.Duration(rand.Int63n(int64(jitterWaitInterval)) - int64(jitterWaitInterval)/2)
+	nextWaitInterval -= time.Duration(rand.Int63n(int64(jitterWaitInterval) / 2))
 
 	return nextWaitInterval
 }
