@@ -2,11 +2,11 @@ package sqlite
 
 import "github.com/mattermost/morph/drivers"
 
-func getDefaultConfig() *Config {
-	return &Config{
+func getDefaultConfig() *driverConfig {
+	return &driverConfig{
 		Config: drivers.Config{
 			MigrationsTable:        "db_migrations",
-			StatementTimeoutInSecs: 60,
+			StatementTimeoutInSecs: 300,
 			MigrationMaxSize:       defaultMigrationMaxSize,
 		},
 	}
